Use f32 floats instead of f64 for the shader
This is a tip from Ray Tracing from the Ground Up and should save a bit
of CPU and memory during renders.

+++ src/patterns.rs
@@ -55,7 +55,7 @@ pub mod tests {
         }
 
         fn pattern_at(&self, Point { x, y, z }: Point) -> Color {
-            Color::new(x, y, z)
+            Color::new(x as f32, y as f32, z as f32)
         }
 
         fn transform(&self) -> Matrix4 {
